Fuzz: move afl fuzz code to the crate
