+++ td-uefi-pi/fuzz/Cargo.toml
@@ -10,7 +10,8 @@ edition = "2018"
 cargo-fuzz = true
 
 [dependencies]
-libfuzzer-sys = "0.4"
+libfuzzer-sys = {version = "0.4", optional = true }
+afl = {version = "*", optional = true }
 
 [dependencies.td-uefi-pi]
 path = ".."
@@ -19,6 +20,9 @@ path = ".."
 [workspace]
 members = ["."]
 
+[features]
+default = ["libfuzzer-sys"]
+fuzz = ["afl"]
 
 [[bin]]
 name = "fv_parser"
@@ -29,5 +33,17 @@ doc = false
 [[bin]]
 name = "hob_parser"
 path = "fuzz_targets/hob_parser.rs"
+test = false
+doc = false
+
+[[bin]]
+name = "afl_hob_parser"
+path = "fuzz_targets/afl_hob_parser.rs"
+test = false
+doc = false
+
+[[bin]]
+name = "afl_fv_parser"
+path = "fuzz_targets/afl_fv_parser.rs"
 test = false
 doc = false
